requirements.txt
requests
beautifulsoup4
pandas
vllm
langchain
faiss-cpu
streamlit
python-dotenv
openai